Remove templates.go file.
