+++ politeiawww/legacy/cmsemail.go
@@ -151,3 +151,138 @@ func (p *LegacyPoliteiawww) emailInvoiceNewComment(userEmail string) error {
 
 	return p.mail.SendTo(subject, body, recipients)
 }
+
+// User CMS invite - Send to user being invited
+type userCMSInvite struct {
+	Email string // User email
+	Link  string // Registration link
+}
+
+const userCMSInviteText = `
+You are invited to join Decred as a contractor! To complete your registration, you will need to use the following link and register on the CMS site:
+
+{{.Link}}
+
+You are receiving this email because {{.Email}} was used to be invited to Decred's Contractor Management System.
+If you do not recognize this, please ignore this email.
+`
+
+var userCMSInviteTmpl = template.Must(
+	template.New("userCMSInvite").Parse(userCMSInviteText))
+
+// User DCC approved - Send to approved user
+type userDCCApproved struct {
+	Email string // User email
+}
+
+const userDCCApprovedText = `
+Congratulations! Your Decred Contractor Clearance Proposal has been approved! 
+
+You are now a fully registered contractor and may now submit invoices.  You should also be receiving an invitation to the contractors room on matrix.  
+If you have any questions please feel free to ask them there.
+
+You are receiving this email because {{.Email}} was used to be invited to Decred's Contractor Management System.
+If you do not recognize this, please ignore this email.
+`
+
+var userDCCApprovedTmpl = template.Must(
+	template.New("userDCCApproved").Parse(userDCCApprovedText))
+
+// DCC submitted - Send to admins
+type dccSubmitted struct {
+	Link string // DCC gui link
+}
+
+const dccSubmittedText = `
+A new DCC has been submitted.
+
+{{.Link}}
+
+Regards,
+Contractor Management System
+`
+
+var dccSubmittedTmpl = template.Must(
+	template.New("dccSubmitted").Parse(dccSubmittedText))
+
+// DCC support/oppose - Send to admins
+type dccSupportOppose struct {
+	Link string // DCC gui link
+}
+
+const dccSupportOpposeText = `
+A DCC has received new support or opposition.
+
+{{.Link}}
+
+Regards,
+Contractor Management System
+`
+
+var dccSupportOpposeTmpl = template.Must(
+	template.New("dccSupportOppose").Parse(dccSupportOpposeText))
+
+// Invoice status update - Send to invoice owner
+type invoiceStatusUpdate struct {
+	Token string // Invoice token
+}
+
+const invoiceStatusUpdateText = `
+An invoice's status has been updated, please login to cms.decred.org to review the changes.
+
+Updated Invoice Token: {{.Token}}
+
+Regards,
+Contractor Management System
+`
+
+var invoiceStatusUpdateTmpl = template.Must(
+	template.New("invoiceStatusUpdate").Parse(invoiceStatusUpdateText))
+
+// Invoice notifications
+var invoiceFirstNotificationTmpl = template.Must(
+	template.New("first_invoice_notification").Parse(invoiceFirstText))
+var invoiceSecondNotificationTmpl = template.Must(
+	template.New("second_invoice_notification").Parse(invoiceSecondText))
+var invoiceFinalNotificationTmpl = template.Must(
+	template.New("final_invoice_notification").Parse(invoiceFinalText))
+
+type invoiceNotification struct {
+	Username string
+	Month    string
+	Year     int
+}
+
+const invoiceFirstText = `
+{{.Username}},
+
+Please submit your invoice for {{.Month}} {{.Year}}.
+
+Regards,
+Contractor Management System
+`
+
+const invoiceSecondText = `
+{{.Username}},
+
+You have not yet submitted an invoice for {{.Month}} {{.Year}}.
+
+Regards,
+Contractor Management System`
+
+const invoiceFinalText = `
+{{.Username}},
+
+You have not yet submitted an invoice for {{.Month}} {{.Year}}.  This is the final warning you will receive, if you delay further, you may not be included in this month's payout.
+
+Regards,
+Contractor Management System
+`
+
+// Invoice new comment - Send to invoice owner
+const invoiceNewCommentText = `
+An administrator has submitted a new comment to your invoice, please login to cms.decred.org to view the message.
+`
+
+var invoiceNewCommentTmpl = template.Must(
+	template.New("invoiceNewComment").Parse(invoiceNewCommentText))
