+++ politeiawww/legacy/useremail.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2020 The Decred developers
+// Copyright (c) 2017-2020 The Decred developers
 // Use of this source code is governed by an ISC
 // license that can be found in the LICENSE file.
 
@@ -169,3 +169,110 @@ func createBody(tpl *template.Template, tplData interface{}) (string, error) {
 
 	return buf.String(), nil
 }
+
+// User email verify - Send verification link to new user
+type userEmailVerify struct {
+	Username string // User username
+	Link     string // Verification link
+}
+
+const userEmailVerifyText = `
+Thanks for joining Politeia, {{.Username}}!
+
+Click the link below to verify your email and complete your registration.
+
+{{.Link}}
+
+You are receiving this notification because this email address was used to
+register a Politeia account.  If you did not perform this action, please ignore
+this email.
+`
+
+var userEmailVerifyTmpl = template.Must(
+	template.New("userEmailVerify").Parse(userEmailVerifyText))
+
+// User key update - Send key verification link to user
+type userKeyUpdate struct {
+	PublicKey string // User new public key
+	Username  string
+	Link      string // Verify key link
+}
+
+const userKeyUpdateText = `
+Click the link below to verify your new identity:
+
+{{.Link}}
+
+You are receiving this notification because a new identity was generated for
+{{.Username}} on Politeia with the following public key. 
+
+Public key: {{.PublicKey}} 
+
+If you did not perform this action, please contact a Politeia administrators in
+the Politeia channel on Matrix.
+
+https://chat.decred.org/#/room/#politeia:decred.org
+`
+
+var userKeyUpdateTmpl = template.Must(
+	template.New("userKeyUpdate").Parse(userKeyUpdateText))
+
+// User password reset - Send password reset link to user
+type userPasswordReset struct {
+	Link string // Password reset link
+}
+
+const userPasswordResetText = `
+Click the link below to continue resetting your password:
+
+{{.Link}}
+
+A password reset was initiated for this Politeia account.  If you did not
+perform this action, it's possible that your account has been compromised.
+Please contact a Politeia administrator in the Politeia channel on Matrix.
+
+https://chat.decred.org/#/room/#politeia:decred.org
+`
+
+var userPasswordResetTmpl = template.Must(
+	template.New("userPasswordReset").Parse(userPasswordResetText))
+
+// User account locked - Send reset password link to user
+type userAccountLocked struct {
+	Link     string // Reset password link
+	Username string
+}
+
+const userAccountLockedText = `
+The Politeia account for {{.Username}} was locked due to too many login
+attempts. You need to reset your password in order to unlock your account:
+
+{{.Link}}
+
+If these login attempts were not made by you, please notify a Politeia
+administrators in the Politeia channel on Matrix.
+
+https://chat.decred.org/#/room/#politeia:decred.org
+`
+
+var userAccountLockedTmpl = template.Must(
+	template.New("userAccountLocked").Parse(userAccountLockedText))
+
+// User password changed - Send to user
+type userPasswordChanged struct {
+	Username string
+}
+
+const userPasswordChangedText = `
+The password has been changed for your Politeia account with the username
+{{.Username}}. 
+
+If you did not perform this action, it's possible that your account has been
+compromised.  Please contact a Politeia administrator in the Politeia channel
+on Matrix.
+
+https://chat.decred.org/#/room/#politeia:decred.org
+`
+
+var userPasswordChangedTmpl = template.Must(
+	template.New("userPasswordChanged").Parse(userPasswordChangedText))
